Update dev dependencies in composer.json
This commit updates the 'require-dev' field in composer.json to match current project requirements. Unused packages were removed and new necessary ones were added. This change ensures all the packages used in the development environment are up-to-date and relevant, and it helps to keep the project maintainable in the long term.